requirements.txt
asgiref==3.7.2
confusable-homoglyphs==3.2.0
Django==3.2.5
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-registration-redux==2.13
django-simple-captcha==0.5.20
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.12.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0